+++ CHANGELOG.md
@@ -14,6 +14,8 @@ Types of changes are:
 
 ## [Unreleased]
 ### Added
+* Resolve local and remote file-system references in JSONSchema
+  documents.
 * Class-generation based on singular types.
 * Class-generation based on combinations of
     - `boolean`